requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
cffi==1.5.0               # via jack-client
click==6.2                # via pip-tools
cycler==0.9.0             # via matplotlib
cython==0.23.4            # via h5py
decorator==4.0.6          # via ipython, traitlets
first==2.0.1              # via pip-tools
h5py==2.5.0
ipdb==0.8.1
ipython-genutils==0.1.0   # via traitlets
ipython==4.0.2
jack-client==0.4.0
matplotlib==1.5.1
numpy==1.10.4
path.py==8.1.2            # via pickleshare
pexpect==4.0.1            # via ipython
pickleshare==0.5          # via ipython
pip-tools==1.4.4
ptyprocess==0.5           # via pexpect
pycparser==2.14           # via cffi
pyparsing==2.0.7          # via matplotlib
pyqtgraph==0.9.10
pyside==1.2.4
python-dateutil==2.4.2    # via matplotlib
pytz==2015.7              # via matplotlib
scipy==0.16.1
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via cycler, h5py, pip-tools, python-dateutil
traitlets==4.1.0          # via ipython
